Expected type of expression was made to depend on type of variable.
At the assignment statement checking that variable's type corresponds
to whatever expression was replaced with mandating the expression type
to be equal to the type of variable.

This allowed to use polymorphic expressions in assignments and those
expressions get monomorphic depending on the type of target variable.
This allowed to say `Int. "x" |= C 2` without saying `C (the Int 2)`.

Thus, we get rid of specialized variant of `C` for `Int`s.

This led to inability to relax expressions when adding an arbitrary
entry to the context since is could change the type of `V n` expression.
Thus, to relax function was removed.
To restore it we need to add a limitation that added entry to the
context does not change any type of any variable used in the expression.
I'm yet too lazy to implement this, thus signatures of define-and-assign
functions got more complicated (and allow to use unassigned variable in
assignment expression, which was prohibited earlier).