`Gen` was redesigned to consider the uniform distribution separately.
During the standard operations (`map`, applicative's application,
monadic bind, alternative's composition) the uniform distribution is
kept as long as possible.

Also, generally `Gen` was made to fail to generate any of the output.
Maybe, it's not the right thing to do, but allows us to filter uniform
`Gen`s nicely and generally implement the `Alternative`.