Thw whole design of lazy container was changed.
The main emphasis was made to the any-sized version (i.e., list
instead of vector). Record was used to such lists instead of a
dependent pair. Potential repetition of runtime length variables
and `Lazy` markup was moved inside the `LzList` type and `LzVect`
was defined mutually with `LzList`.

Also, new constructors were added to the `LzVect` type to construct
from eager lists effectively and to support lazy applicative ops.