Monadic composition of `Gen`s was made to produce `Raw` always.
This is done because it is impossible to preserve truely lazy uniform
list during monadic composition of uniform `Gen`s because sizes of
underlying lists are known only when forcing the LHS of `>>=`.
This means that each `>>=` composition forces the left uniform `Gen`
(if the left one if uniform) to calculate each of its element.