Element-replication lazier list generator was added as a primitive.
This allows to save memory by not having a list of some length.
Also, this allows to index faster than "eager of replicate" variant.
But what is the most important, this variant of list interacts and
combines with other variants of lists, thus can be used to model
non-uniform stuff using the very same lazier list data structure.