`oneOf` was changed to take only non-empty stuff.
Since it's mainly used with literals, no one should even notice the
change. Old one taking `List` is not called `oneOf'`.

This change is done to get rid of an `empty` at the end of `choiceMap`
application (since `List`s can be potentially empty and have the empty
as a base case).