[ refine ] Separate error reporting for implicit and explicit givens
Well, this change has both pros and contras.
Pros are that use will known what givens list (for implicits or explicits)
is problematic. Contras are that these lists are now checked one after
another, so first errors will be shown only for the implicits list, then
only for the explicits list.